Add Get Started in docs (#279)
* Add Get Started in docs.

Going through the docs for the first time, I couldn't get it to work. I was missing the schema file,
the bean had a dependency that wasn't necessary clear (unless you pull down the repository and inspect).

So this change adds a new section in the documentation which enables anyone to get started with GraphQL
and Micronaut. The step by step guide goes through lightweight app creation or configuration, creating
the schema in the `resources` folder, creating the high-level graphql bean, and launching the app to
do some testing.

* Use the code we already have

* Cleanup

* Update quickStart.adoc

Co-authored-by: Michaël Ortali <michael.ortali@gmail.com>