Ensure reactive context is propagated
DefaultGraphQLInvocation and DefaultGraphQLExecutionInputCustomizer are updated
to use Mono and Flux in their internal implementation in order to ensure that
the Micronaut PropagationContext is carried appropriately through the invocation
flow. Previous use of the bare io.micronaut.core.async.publisher.Publishers API
in DefaultGraphQLExecutionInputCustomizer was causing the context not to be
propagated as desired when including the Micrometer Context Propagation library.

A test is added to verify the context propagation works as expected.

Some additional cleanup is done throughout the test suite to reduce the scope
of included Micronaut beans to the specific tests in order to make it easier to
test different setups.

Resolves #495